# documentation/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Golden Lift API",
    "version": "0.4.0",
    "description": "Implemented B1-B4 and Dynamic Catalog Core: anonymous browsing/product projection, live staff Identity, complete category administration, dynamic product types/attributes/groups/units and guarded schema/product workflows. Dynamic operations require reviewed Catalog v1.2 cutover. Media delivery, full product/technical editing UX and Inquiry workflows remain later milestones."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/staff/admins/{id}": {
      "get": {
        "operationId": "getAdmin",
        "description": "Super Admin caller and ADMIN target only.",
        "security": [
          {
            "staffSession": []
          },
          {
            "staffLocalSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffAccount"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "editAdmin",
        "description": "Super Admin caller and ADMIN target only; email changes revoke credentials.",
        "security": [
          {
            "staffSession": []
          },
          {
            "staffLocalSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must exactly match an approved staff origin."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActionToken"
            },
            "description": "Read csrfToken from login/current-session JSON; it belongs to the current cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminEditInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffAccount"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAdmin",
        "description": "Super Admin only; soft deletion retains records and email uniqueness and revokes credentials.",
        "security": [
          {
            "staffSession": []
          },
          {
            "staffLocalSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must exactly match an approved staff origin."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActionToken"
            },
            "description": "Read csrfToken from login/current-session JSON; it belongs to the current cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpectedVersionInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffAccount"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/categories": {
      "post": {
        "operationId": "createCategory",
        "description": "ADMIN only. SUPER_ADMIN cannot manage content.",
        "security": [
          {
            "staffSession": []
          },
          {
            "staffLocalSession": []
          }
        ],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must exactly match an approved staff origin."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ActionToken"
            },
            "description": "Read csrfToken from login/current-session JSON; it belongs to the current cookie."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Safe error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listAdminCategories",
        "description": "ADMIN child list (roots by default). Stable sortOrder/UUID order. Cursor binds parent, locale, list revision and purpose; stale cursor is 409. No complete-tree response; load children on demand. Up to 100 children per page. Saved translations and active counts are private.",
        "security": [
          {
            "staffSession": []
          },
          {
            "staffLocalSession": []
          }
        ],
        "parameters": [
          {
            "name": "parentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Locale"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 1024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryCollectionPage"
                }
              }
            }
          },
          "400": {
            "description": "Malformed input or cursor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, revoked, disabled or deleted session/account.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Requires ADMIN; SUPER_ADMIN is excluded. Unsafe Origin/CSRF is denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Selected category, parent or route is unavailable/deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Stale row, list, path or preview precondition. Reload before retrying.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "JSON body exceeds 64 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Invalid destination, membership, cover, hierarchy or ordering-gap recovery beyond the documented bound.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Identity/Catalog is unavailable or the bounded transaction retry budget is exhausted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Locale": {
        "type": "string",
        "enum": ["ar", "en", "ckb"],
        "default": "ar"
      },
      "Version": {
        "type": "string",
        "pattern": "^[1-9][0-9]{0,18}$",
        "description": "Positive signed bigint represented without JavaScript precision loss."
      },
...
      "ApiError": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "requestId"],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
      "Category": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "parentId",
          "name",
          "description",
          "slug",
          "locale",
          "resolvedNameLocale",
          "sortOrder",
          "version"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          },
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "resolvedNameLocale": {
            "$ref": "#/components/schemas/Locale"
          },
          "sortOrder": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
...
      "StaffAccount": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "email", "displayName", "role", "status", "version", "createdAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "role": {
            "enum": ["ADMIN", "SUPER_ADMIN"]
          },
          "status": {
            "enum": ["INVITED", "ACTIVE", "DISABLED"]
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ActionToken": {
        "type": "string",
        "pattern": "^[A-Za-z0-9_-]{43}$"
      },
...
      "AdminEditInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expectedVersion"],
        "properties": {
          "expectedVersion": {
            "$ref": "#/components/schemas/Version"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          }
        },
        "description": "Supply email or displayName. Roles and status cannot be assigned through this route."
      },
      "ExpectedVersionInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["expectedVersion"],
        "properties": {
          "expectedVersion": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
      "CategoryTranslation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["locale", "name"],
        "properties": {
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "description": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          }
        }
      },
      "CategoryCreateInput": {
        "type": "object",
        "additionalProperties": false,
        "required": ["translations"],
        "properties": {
          "parentId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "expectedParentVersion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Version"
              },
              {
                "type": "null"
              }
            ]
          },
          "translations": {
            "type": "array",
            "minItems": 1,
            "maxItems": 3,
            "items": {
              "$ref": "#/components/schemas/CategoryTranslation"
            },
            "description": "Unique locales; an Arabic name is mandatory."
          },
          "coverAssetId": {
            "type": ["string", "null"],
            "format": "uuid"
          }
        },
        "description": "Root creation omits parent fields or supplies null; child creation requires both the parent UUID and current parent version. Optional coverAssetId must refer to an already verified active image registration; null clears a cover, omission preserves it on edit. Parent/order changes are rejected on generic PATCH."
      },
...
      "CategoryListRevision": {
        "type": "string",
        "pattern": "^l1-[a-f0-9]{64}$",
        "description": "Opaque scope-bound state precondition. Re-read after 409; never an authorization credential."
      },
...
      "StoredCategoryTranslation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["locale", "name", "description", "slug", "version"],
        "properties": {
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "description": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
      "AdminCategory": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "parentId",
          "name",
          "description",
          "slug",
          "locale",
          "resolvedNameLocale",
          "sortOrder",
          "version",
          "coverAssetId",
          "translations",
          "missingTranslationLocales",
          "activeChildCount",
          "activeProductCount",
          "canAddChildren",
          "canAddProducts"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": ["string", "null"]
          },
          "slug": {
            "type": ["string", "null"]
          },
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "resolvedNameLocale": {
            "$ref": "#/components/schemas/Locale"
          },
          "sortOrder": {
            "type": "string",
            "pattern": "^-?[0-9]+$"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "coverAssetId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StoredCategoryTranslation"
            }
          },
          "missingTranslationLocales": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Locale"
            }
          },
          "activeChildCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "activeProductCount": {
            "type": "string",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "canAddChildren": {
            "type": "boolean"
          },
          "canAddProducts": {
            "type": "boolean"
          }
        },
        "description": "ADMIN only. Display fields fall back to Arabic per field; translations contain actual saved values. Missing locales are explicit. Eligibility derives from active child/product counts; there is no stored mode."
      },
      "CategoryCollectionPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["parentId", "listRevision", "items", "nextCursor"],
        "properties": {
          "parentId": {
            "type": ["string", "null"],
            "format": "uuid"
          },
          "listRevision": {
            "$ref": "#/components/schemas/CategoryListRevision"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCategory"
            }
          },
          "nextCursor": {
            "type": ["string", "null"],
            "maxLength": 1024
          }
        }
      },
...
    },
    "securitySchemes": {
      "staffSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-gl_staff",
        "description": "Production HttpOnly, Secure, SameSite=Strict cookie. All protected mutations also require the session-bound X-CSRF-Token and approved Origin."
      },
      "staffLocalSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gl_staff",
        "description": "Development-only cookie, never used in production."
      }
    }
  }
}